+++ src/thor/timedep_reverse.cc
@@ -3,7 +3,6 @@
 #include "midgard/logging.h"
 #include "thor/timedep.h"
 #include <algorithm>
-#include <map>
 #include <string>
 
 using namespace valhalla::baldr;
@@ -19,7 +18,7 @@ constexpr uint64_t kInitialEdgeLabelCount = 500000;
 constexpr uint32_t kMaxIterationsWithoutConvergence = 800000;
 
 // Default constructor
-TimeDepReverse::TimeDepReverse() : AStarPathAlgorithm() {
+TimeDepReverse::TimeDepReverse() : TimeDepForward() {
   mode_ = TravelMode::kDrive;
   travel_type_ = 0;
   adjacencylist_ = nullptr;
@@ -33,7 +32,7 @@ TimeDepReverse::~TimeDepReverse() {
 }
 
 void TimeDepReverse::Clear() {
-  AStarPathAlgorithm::Clear();
+  TimeDepForward::Clear();
   edgelabels_rev_.clear();
 }